Working on Town Blockout
My scene is based on the painting I did a while ago so that's what I'm trying to recreate as the environment at the moment. 
This commit is me learning about BSP Brushes to blockout a scene as well as bookmarks to switch between screenshot places in order to record process shots of the creation of the scene.
Actual Changes:
-Not much, just messing around with level blockout.